chore: #201 사용자의 활동 통계 및 작성한 인증 목록 전체 조회 API 버저닝 도입
